A lot of CHANGES, please take some time and read these: -Moved the log4j.properties file to the resource folder so that it is already included in our classpath. -Changed the logging setting to only show warn and removed all of hibernate SQL logging also. -Changed all the insert_script and added a few more. -Changed the dump file for the database to a freshless build database with good data. -Fixed some hibernate properties -Changed the log file directory value in CerberusServerBootStrap -Changed the SocketId from Integer to Long, and changed the database from integer to varchar(10) and unique. -Changed all the Messages to keep the RFID in string and Socket in Long -Fixed all the test cases in JUnit -Added the ClientDataGenerator to generate good message data from the sample data in the database. (Can add later to the arrays) -Added a log of stuff to UserWorkflow and UserService when I was debugging. -Fixed the User and Login bean to deal with the recursion problem later -THAT'S IT

+++ cerberus-server/src/main/java/com/cerberus/server/message/CurrentConsumptionMessage.java
@@ -9,7 +9,8 @@ public class CurrentConsumptionMessage extends Message {
 	private final String rfidNumber;
 
 	@JsonCreator
-	public CurrentConsumptionMessage(@JsonProperty("socketId") int socketId, @JsonProperty("timestamp") long timestamp,
+	public CurrentConsumptionMessage(@JsonProperty("socketId") long socketId,
+			@JsonProperty("timestamp") long timestamp,
 			@JsonProperty("current") int current, @JsonProperty("rfidNumber") String rfidNumber) {
 		super(MessageType.CURRENT, socketId, timestamp);
 		this.current = current;
