+++ cerberus-server/src/main/java/com/cerberus/server/workflow/OutletWorkflow.java
@@ -64,7 +64,7 @@ public class OutletWorkflow extends Workflow {
 		return serviceFactory.getOutletService().getSocketBySocketId(id);
 	}
 
-	public Socket getSocketBySerialNumber(Integer serialNumber) {
+	public Socket getSocketBySerialNumber(Long serialNumber) {
 		return serviceFactory.getOutletService().getSocketBySerialNumber(serialNumber);
 	}
