+++ cerberus-server/src/test/java/com/cerberus/server/message/RFIDAuthResponseMessageTest.java
@@ -1,22 +1,21 @@
 package com.cerberus.server.message;
 
-import static org.junit.Assert.assertThat;
-import static org.junit.Assert.assertTrue;
-
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
+import static org.junit.Assert.assertThat;
+import static org.junit.Assert.assertTrue;
 
 import org.junit.Before;
 import org.junit.Test;
 
 public class RFIDAuthResponseMessageTest {
 
-	private int socketId;
+	private long socketId;
 	private long timestamp;
 
 	@Before
 	public void setup() {
-		socketId = 12345;
+		socketId = 12345L;
 		timestamp = System.currentTimeMillis();
 	}
